Added e2ee sanity tests
